+++ docs/config/index.html
@@ -297,50 +297,8 @@
   
 </li>
         
-          <li class="md-nav__item">
-  <a href="#port" title="port" class="md-nav__link">
-    port
-  </a>
-  
-</li>
-        
-          <li class="md-nav__item">
-  <a href="#interval" title="interval" class="md-nav__link">
-    interval
-  </a>
-  
-</li>
-        
-          <li class="md-nav__item">
-  <a href="#serve" title="serve" class="md-nav__link">
-    serve
-  </a>
-  
-</li>
-        
-          <li class="md-nav__item">
-  <a href="#watchexps" title="watchexps" class="md-nav__link">
-    watchexps
-  </a>
-  
-</li>
-        
-          <li class="md-nav__item">
-  <a href="#quiet" title="quiet" class="md-nav__link">
-    quiet
-  </a>
-  
-</li>
-        
       </ul>
     </nav>
-  
-</li>
-      
-        <li class="md-nav__item">
-  <a href="#light-server" title="light-server" class="md-nav__link">
-    light-server
-  </a>
   
 </li>
       
@@ -439,50 +397,8 @@
   
 </li>
         
-          <li class="md-nav__item">
-  <a href="#port" title="port" class="md-nav__link">
-    port
-  </a>
-  
-</li>
-        
-          <li class="md-nav__item">
-  <a href="#interval" title="interval" class="md-nav__link">
-    interval
-  </a>
-  
-</li>
-        
-          <li class="md-nav__item">
-  <a href="#serve" title="serve" class="md-nav__link">
-    serve
-  </a>
-  
-</li>
-        
-          <li class="md-nav__item">
-  <a href="#watchexps" title="watchexps" class="md-nav__link">
-    watchexps
-  </a>
-  
-</li>
-        
-          <li class="md-nav__item">
-  <a href="#quiet" title="quiet" class="md-nav__link">
-    quiet
-  </a>
-  
-</li>
-        
       </ul>
     </nav>
-  
-</li>
-      
-        <li class="md-nav__item">
-  <a href="#light-server" title="light-server" class="md-nav__link">
-    light-server
-  </a>
   
 </li>
       
@@ -515,14 +431,7 @@
       &quot;font&quot;
     ],
     &quot;external&quot;: []
-  },
-  &quot;port&quot;: 5550,
-  &quot;interval&quot;: 750,
-  &quot;serve&quot;: &quot;build&quot;,
-  &quot;watchexps&quot;: [
-    &quot;game/** # npm run build&quot;
-  ],
-  &quot;quiet&quot;: true
+  }
 }
 </code></pre>
 
@@ -543,21 +452,6 @@
 <p><em>In the default config above, the directories <strong>data</strong> and <strong>font</strong> are added to the project as additional entries. If you don't need that functionality, you can remove the directories, as well as the entries in the <strong>pnkit.json</strong> file.</em> </p>
 <h4 id="external">external</h4>
 <p>If you would like to include directories in the build from outside of the <em>game</em> directory, you can add them to this array. The steps are the same as for the <code>game</code> key above. </p>
-<h3 id="port">port</h3>
-<p>The port the watch server will use. The default port is 5550.</p>
-<h3 id="interval">interval</h3>
-<p>The time in milliseconds between watch updates. The default is 750.</p>
-<h3 id="serve">serve</h3>
-<p>The directory that will be served up. The default is the <em>build</em> directory.</p>
-<h3 id="watchexps">watchexps</h3>
-<p>The server watch expressions. </p>
-<p>The default is to watch the <em>game</em> directory and run a "clean-n-build" on changes.</p>
-<p><em>In the config above, the pound sign (#) is not a comment, but a separator. Do not remove it. For more info see the <strong>light-server</strong> entry below.</em></p>
-<h3 id="quiet">quiet</h3>
-<p>Keep the noise down on the watch server output. The default is <code>true</code>.</p>
-<hr />
-<h2 id="light-server">light-server</h2>
-<p>Phaser Node Kit uses the crafty <strong>light-server</strong> package to serve up your game in progress. You can learn more about its usage options at <a href="https://www.npmjs.com/package/light-server">https://www.npmjs.com/package/light-server</a>.</p>
