+++ docs/sitemap.xml
@@ -4,7 +4,7 @@
     
     <url>
      <loc>/</loc>
-     <lastmod>2017-04-11</lastmod>
+     <lastmod>2017-04-12</lastmod>
      <changefreq>daily</changefreq>
     </url>
     
@@ -12,7 +12,7 @@
     
     <url>
      <loc>/install/</loc>
-     <lastmod>2017-04-11</lastmod>
+     <lastmod>2017-04-12</lastmod>
      <changefreq>daily</changefreq>
     </url>
     
@@ -20,7 +20,7 @@
     
     <url>
      <loc>/cli/</loc>
-     <lastmod>2017-04-11</lastmod>
+     <lastmod>2017-04-12</lastmod>
      <changefreq>daily</changefreq>
     </url>
     
@@ -28,7 +28,7 @@
     
     <url>
      <loc>/workflow/</loc>
-     <lastmod>2017-04-11</lastmod>
+     <lastmod>2017-04-12</lastmod>
      <changefreq>daily</changefreq>
     </url>
     
@@ -36,7 +36,7 @@
     
     <url>
      <loc>/config/</loc>
-     <lastmod>2017-04-11</lastmod>
+     <lastmod>2017-04-12</lastmod>
      <changefreq>daily</changefreq>
     </url>
     
@@ -44,7 +44,7 @@
     
     <url>
      <loc>/notes/</loc>
-     <lastmod>2017-04-11</lastmod>
+     <lastmod>2017-04-12</lastmod>
      <changefreq>daily</changefreq>
     </url>
     
@@ -52,7 +52,7 @@
     
     <url>
      <loc>/issues/</loc>
-     <lastmod>2017-04-11</lastmod>
+     <lastmod>2017-04-12</lastmod>
      <changefreq>daily</changefreq>
     </url>
