+++ lib/build.js
@@ -24,7 +24,7 @@ class Build {
     })
   }
 
-  run() {
+  run(refresh=true) {
 
     p.log('Generating Build')
 
@@ -74,7 +74,9 @@ class Build {
       } else {
         let d = new Date()
         bundler.bundle().then(() => {
-          this.refresh()
+          if (refresh === true) {
+            this.refresh()
+          }
           p.ok(`Build Completed @${d.toLocaleTimeString()}`) 
         })
       }
