+++ lib/clean.js
@@ -22,7 +22,7 @@ class Clean {
 
     //rebuild
     let builder = new Build()
-    builder.run()
+    builder.run(false)
   }
 }
