+++ tpl/game/js/Engine.js
@@ -1,0 +1,20 @@
+
+class Engine {
+  constructor(game) {
+    this.game = game
+  }
+
+  run() {
+    let node = this.game.add.image(
+      this.game.world.centerX, 
+      this.game.world.centerY, 
+      'node')
+
+    node.anchor.set(.5)
+  }
+
+  update() { }
+  render() { }
+}
+
+module.exports = Engine
