+++ tpl/game/js/states/PreloadState.js
@@ -1,9 +1,7 @@
 
 /* Preload State */
 class PreloadState {
-
   preload() {
-
     this.preloadBar = null
  
     this.preloadBar = this.game.add.sprite(
@@ -24,7 +22,6 @@ class PreloadState {
   }
 
   update() { }
-
   render() { }
 }
